- Added Maslovian needs to PopulationAgent - Changed prices to int - Changed cash, etc. to int - Non stable build

+++ Assets/Actions/ActionContext.cs
@@ -32,8 +32,8 @@ public class ActionContext
 
     //Submit Bid/Ask
     public Good GoodToSubmit;
-    public decimal BidToSubmit;
-    public decimal AskToSubmit;
+    public int BidToSubmit;
+    public int AskToSubmit;
 
     public LemonadeStandResultObject HasSubmittingCompany()
     {
